# backend/docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ALPACART API",
    "description": "API Backend de Alpacart — Marca premium de alpaca peruana",
    "version": "1.0.0",
    "contact": {
      "name": "Alpacart Textiles",
      "url": "https://alpacart.com"
    }
  },
  "servers": [
    { "url": "http://localhost:8000/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registro de cliente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "mateo.q@alpacart.com" },
                  "firstName": { "type": "string", "minLength": 2, "example": "Mateo" },
                  "lastName": { "type": "string", "minLength": 2, "example": "Quispe" },
                  "password": { "type": "string", "minLength": 6, "example": "securePassword123" }
                },
                "required": ["email", "firstName", "lastName", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Cliente registrado exitosamente" },
          "409": { "description": "Email ya registrado" }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": ["Customers"],
        "summary": "Mi carrito",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Carrito con items" } }
      },
      "delete": {
        "tags": ["Customers"],
        "summary": "Vaciar carrito",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Carrito vaciado" } }
      }
    },
    "/cart/items": {
      "post": {
        "tags": ["Customers"],
        "summary": "Agregar al carrito",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": { "type": "string", "format": "uuid" },
                  "variantId": { "type": "string", "format": "uuid" },
                  "quantity": { "type": "integer", "minimum": 1 }
                },
                "required": ["productId", "quantity"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Item agregado" } }
      }
    },
    "/cart/items/{id}": {
      "patch": {
        "tags": ["Customers"],
        "summary": "Actualizar cantidad",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cantidad actualizada" } }
      },
      "delete": {
        "tags": ["Customers"],
        "summary": "Eliminar item del carrito",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Item eliminado" } }
      }
    },
    "/checkout": {
      "post": {
        "tags": ["Customers"],
        "summary": "Procesar checkout",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "couponCode": { "type": "string" },
                  "idempotencyKey": { "type": "string", "description": "Idempotency-Key header" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Orden creada" } }
      }
    },
    "/account/profile": {
      "get": {
        "tags": ["Customers"],
        "summary": "Mi perfil",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Perfil del cliente" } }
      },
      "put": {
        "tags": ["Customers"],
        "summary": "Actualizar perfil",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Perfil actualizado" } }
      }
    },
    "/account/password": {
      "put": {
        "tags": ["Customers"],
        "summary": "Cambiar contraseña",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Contraseña actualizada" } }
      }
    },
    "/account/addresses": {
      "get": {
        "tags": ["Customers"],
        "summary": "Mis direcciones",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de direcciones" } }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Crear dirección",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Dirección creada" } }
      }
    },
    "/account/addresses/{id}": {
      "delete": {
        "tags": ["Customers"],
        "summary": "Eliminar dirección",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Dirección eliminada" } }
      }
    },
    "/wishlist": {
      "get": {
        "tags": ["Customers"],
        "summary": "Lista de deseos",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de deseos" } }
      }
    },
    "/wishlist/items": {
      "post": {
        "tags": ["Customers"],
        "summary": "Agregar/quitar de wishlist",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Item agregado/removido" } }
      }
    },
...
    "/coupons": {
      "get": {
        "tags": ["Marketing"],
        "summary": "Listar cupones",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de cupones" } }
      },
      "post": {
        "tags": ["Marketing"],
        "summary": "Crear cupón",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": { "type": "string", "example": "WELCOME10" },
                  "type": { "type": "string", "example": "percentage" },
                  "value": { "type": "number", "example": 10 }
                },
                "required": ["code", "type", "value"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Cupón creado" } }
      }
    },
    "/coupons/{id}": {
      "get": {
        "tags": ["Marketing"],
        "summary": "Obtener cupón",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cupón encontrado" } }
      },
      "put": {
        "tags": ["Marketing"],
        "summary": "Actualizar cupón",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cupón actualizado" } }
      },
      "delete": {
        "tags": ["Marketing"],
        "summary": "Eliminar cupón",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cupón eliminado" } }
      }
    },
    "/coupons/validate": {
      "post": {
        "tags": ["Marketing"],
        "summary": "Validar cupón",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": { "type": "string", "example": "WELCOME10" },
                  "cartSubtotal": { "type": "number", "example": 100 }
                },
                "required": ["code"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Resultado de validación" } }
      }
    },
    "/campaigns": {
      "get": {
        "tags": ["Marketing"],
        "summary": "Listar campañas",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de campañas" } }
      },
      "post": {
        "tags": ["Marketing"],
        "summary": "Crear campaña",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Campaña creada" } }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": ["Marketing"],
        "summary": "Obtener campaña",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Campaña encontrada" } }
      },
      "put": {
        "tags": ["Marketing"],
        "summary": "Actualizar campaña",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Campaña actualizada" } }
      },
      "delete": {
        "tags": ["Marketing"],
        "summary": "Eliminar campaña",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Campaña eliminada" } }
      }
    },
    "/promotions": {
      "get": {
        "tags": ["Marketing"],
        "summary": "Listar promociones",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de promociones" } }
      },
      "post": {
        "tags": ["Marketing"],
        "summary": "Crear promoción",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Promoción creada" } }
      }
    },
    "/promotions/{id}": {
      "get": {
        "tags": ["Marketing"],
        "summary": "Obtener promoción",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Promoción encontrada" } }
      },
      "put": {
        "tags": ["Marketing"],
        "summary": "Actualizar promoción",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Promoción actualizada" } }
      },
      "delete": {
        "tags": ["Marketing"],
        "summary": "Eliminar promoción",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Promoción eliminada" } }
      }
    },
...
    "/stock": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Listar stock",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de stock" } }
      }
    },
    "/stock/{id}/adjust": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Ajustar stock",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Stock ajustado" } }
      }
    },
    "/movements": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Listar movimientos",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de movimientos" } }
      }
    },
    "/transfers": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Listar transferencias",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de transferencias" } }
      }
    },
...
  }
}